+++ src/richie/plugins/lti_consumer/models.py
@@ -47,14 +47,16 @@ class LTIConsumer(CMSPlugin):
         - context_id from current site domain
         - url if predefined provider is used
         """
-        if not self.id:
-            if self.lti_provider and not self.url:
-                if self.lti_provider.get("is_base_url_regex"):
-                    self.url = "{:s}".format(
-                        exrex.getone(self.lti_provider.get("base_url", "")),
-                    )
-                else:
-                    self.url = self.lti_provider.get("base_url", "")
+        lti_provider = self.lti_provider
+
+        if lti_provider and not self.url:
+            is_regex = lti_provider.get("is_base_url_regex")
+            base_url = lti_provider.get("base_url", "")
+            if is_regex:
+                self.url = "{:s}".format(exrex.getone(base_url))
+            else:
+                self.url = base_url
+
         super().save(*args, **kwargs)
 
     @property
@@ -87,62 +89,58 @@ class LTIConsumer(CMSPlugin):
         """
         return f"padding-bottom: {self.inline_ratio * 100}%"
 
-    def auth_parameters(self, edit=False):
+    def get_resource_link_id(self):
+        """Use the plugin id as resource_link_id field."""
+        return str(self.id)
+
+    def get_content_parameters(self, edit=False):
         """
-        Builds required parameters for LTI authentication
+        Convenient wrapper to authorize and return required parameters
+        for LTI content consumption
         """
         role = INSTRUCTOR if edit else STUDENT
         site = get_current_site()
-        return {
-            "lti_message_type": self.lti_provider.get(
-                "display_name", self.lti_provider_id
-            ),
+
+        lti_parameters = {
+            "lti_message_type": "basic-lti-launch-request",
             "lti_version": "LTI-1p0",
-            "resource_link_id": str(self.id),
+            "resource_link_id": self.get_resource_link_id(),
             "context_id": site.domain,
             "user_id": "richie",
             "lis_person_contact_email_primary": "",
             "roles": role,
         }
 
-    def authorize(self):
-        """
-        Returns headers from LTI authentication
-        """
-        client = oauth1.Client(
-            client_key=self.lti_provider.get("oauth_consumer_key"),
-            client_secret=self.lti_provider.get("shared_secret"),
+        # Get credentials from the predefined LTI provider if any, or from the model otherwise
+        oauth_consumer_key = (
+            self.lti_provider.get("oauth_consumer_key", "")
+            if self.lti_provider_id
+            else self.oauth_consumer_key
         )
-
+        shared_secret = (
+            self.lti_provider.get("shared_secret", "")
+            if self.lti_provider_id
+            else self.shared_secret
+        )
+        client = oauth1.Client(
+            client_key=oauth_consumer_key, client_secret=shared_secret
+        )
         _uri, headers, _body = client.sign(
             self.url,
             http_method="POST",
-            body=self.auth_parameters(),
+            body=lti_parameters,
             headers={"Content-Type": "application/x-www-form-urlencoded"},
         )
-        return headers
 
-    def build_content_parameters(self, auth_headers, edit=False):
-        """
-        Builds required parameters for LTI content consumption
-        """
         oauth_dict = dict(
             param.strip().replace('"', "").split("=")
-            for param in auth_headers["Authorization"].split(",")
+            for param in headers["Authorization"].split(",")
         )
-        signature = oauth_dict["oauth_signature"]
-        oauth_dict["oauth_signature"] = unquote(signature)
+        oauth_dict["oauth_signature"] = unquote(oauth_dict["oauth_signature"])
         oauth_dict["oauth_nonce"] = oauth_dict.pop("OAuth oauth_nonce")
 
-        oauth_dict.update(self.auth_parameters(edit=edit))
+        oauth_dict.update(lti_parameters)
         return oauth_dict
-
-    def content_parameters(self, edit=False):
-        """
-        Convenient wrapper to authorize and return required parameters
-        for LTI content consumption
-        """
-        return self.build_content_parameters(self.authorize(), edit=edit)
 
     def __str__(self):
         return self.url