+++ tests/plugins/lti_consumer/test_factories.py
@@ -1,7 +1,7 @@
 """
 Factory tests
 """
-from django.test import TestCase
+from django.test import TestCase, override_settings
 
 from richie.plugins.lti_consumer.factories import LTIConsumerFactory
 
@@ -9,7 +9,23 @@ from richie.plugins.lti_consumer.factories import LTIConsumerFactory
 class LTIConsumerFactoriesTestCase(TestCase):
     """Tests for the LTIConsumer factory"""
 
-    def test_factories_lti_consumer_create_success(self):
-        """Factory creation success."""
+    @override_settings(
+        RICHIE_LTI_PROVIDERS={
+            "lti_provider_test": {
+                "base_url": "http://localhost:8060/lti/videos/",
+                "is_base_url_regex": False,
+            }
+        }
+    )
+    def test_factories_lti_consumer_create_with_lti_provider(self):
+        """
+        The url field should be computed by the model's "save" method if an
+        LTI provider is defined.
+        """
         lti_consumer = LTIConsumerFactory()
+        self.assertIn(lti_consumer.url, "http://localhost:8060/lti/videos/")
+
+    def test_factories_lti_consumer_create_without_lti_provider(self):
+        """The url field should be set to a random value for a custom provider."""
+        lti_consumer = LTIConsumerFactory(lti_provider_id=None)
         self.assertIsNotNone(lti_consumer.url)